Split Conversion from Size into own module
